requirements.txt
Flask==0.12
python-telegram-bot==5.3.0
requests==2.13.0
giphypop==0.3
